pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-collab"
version = "0.3.0"
description = "Multi-Agent Orchestration Engine — orchestrate Claude Code, Codex, Aider to collaborate on projects"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "JianFei Gan", email = "jfcoder@163.com"}
]
keywords = ["ai", "agent", "claude-code", "codex", "aider", "orchestration", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
agent-collab = "agent_collab.cli:app"

[project.entry-points."agent_collab.plugins"]
# Example: uncomment to register a built-in plugin
# sample-agent = "agent_collab.plugins.interfaces:AgentPlugin"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
